+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## develop branch
 
-Nothing yet.
+## New
+
+* Checks\IsDouble - added
 
 ## 2.12.2 - Sat Sep 19 2015